Fix for deprecated wp_get_sites(); Add category fields to the network admin new site form.
* Replaced wp_get_sites() which is deprecated since 4.6 and calls
get_sites() instead. The new function returns a maximum of 100 sites by
default, so the parameter 'number' => null was added to retrieve all
sites in networks larger than 100 sites;

* Added action for 'network_site_new_form' so the category fields are
also displayed in the network admin new site form;

* The 'add_signup_meta' filter is not fired up when the new site is
created by the network admin, so I had to move the code which gets the
category form fields to the wpmu_new_blog() function. Not so fancy, but
it work for both cases, so I removed (commented out) the action/function
for 'add_signup_meta'.